requirements.txt
annotated-types==0.7.0
anyio==4.7.0
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cookiecutter==2.6.0
distro==1.9.0
Django==5.1.4
django-browser-reload==1.17.0
django-tailwind==3.8.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
jiter==0.8.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.57.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
text-unidecode==1.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.2.3
